# docs/openapi/callpulse-gateways-acl-dialplans.yaml
openapi: 3.0.3
info:
  title: 'FS PBX API v1 — Gateways, Access Controls, Dialplans'
  version: 1.0.0
  description: |
    Contrat des trois ressources ajoutées à l'API v1 de FS PBX pour CallPulse :
    passerelles SIP (`gateways`), listes de contrôle d'accès (`access-controls`)
    et plans de numérotation (`dialplans`).

    Ce document est le contrat de référence : il est écrit avant les contrôleurs
    et fait foi en cas de divergence avec une implémentation.

    ## Authentification

    Toutes les opérations exigent un jeton personnel Sanctum envoyé en
    `Authorization: Bearer <token>`. Deux couches distinctes s'appliquent, dans
    cet ordre, et **elles ne renvoient pas la même forme de corps d'erreur** :

    1. `RequireBearerToken` — en-tête `Authorization` absent → `401`
       `{"success": false, "message": "Bearer token required."}` (forme « middleware »).
    2. `auth:sanctum` — jeton présent mais invalide ou révoqué → `401`
       `{"error": {"type": "authentication_error", ...}}` (forme « enveloppe »).
    3. `AuthorizeUser` — permission manquante → `403` forme « middleware » avec
       `error.code = "forbidden_permission"` ; domaine hors périmètre de
       l'utilisateur → `403` forme « enveloppe » avec `error.code = "forbidden_domain"`.

    Un client doit donc accepter les **deux** formes sur `401` et sur `403`.

    ## Débit

    Limiteur `throttle:api` : **60 requêtes par minute** par utilisateur
...

    ## Codes de retour

    | Code | Signification |
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  '/api/v1/access-controls':
    get:
      tags: ['Access Controls']
      operationId: listAccessControls
      summary: 'Lister les listes de contrôle d''accès de l''instance'
      description: |
        Renvoie toutes les listes d'accès de l'instance, **sans leurs nœuds**
        (utiliser `GET /access-controls/{access_control_uuid}` pour les obtenir).

        Ces listes sont **globales** : elles ne sont pas filtrées par domaine,
        parce que `v_access_controls` ne porte pas de `domain_uuid`. Un appelant
        voit donc les listes de tous les tenants de l'instance.

        La réponse suit l'enveloppe de liste de l'API v1
        (`object`, `url`, `has_more`, `data`). Cette collection n'est pas paginée :
        `has_more` vaut toujours `false`.

        Permission requise : `access_control_view`.
      responses:
        200:
          description: Succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessControlListResponse'
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }
  '/api/v1/access-controls/reload':
    post:
      tags: ['Access Controls']
      operationId: reloadAccessControls
      summary: 'Recharger les listes d''accès dans FreeSWITCH'
      description: |
        Invalide le cache de configuration `acl.conf` puis émet `reloadacl` sur
        l'event socket. Aucune écriture en base.

        Une réponse moteur commençant par `-ERR` — event socket injoignable
        compris — produit un `422` : un rechargement qui n'a pas eu lieu n'est
        jamais renvoyé en succès. C'est ce qui empêche de croire à une liste
        blanche appliquée alors que `iptables` bloque encore le fournisseur.

        Sans corps de requête.

        Note d'implémentation : cette route littérale doit être déclarée **avant**
        `/access-controls/{access_control_uuid}`.

        Permission requise : `access_control_view`.
      responses:
        200:
          description: 'Rechargement effectué.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessControlReloadResponse'
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        422: { $ref: '#/components/responses/SwitchError' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }
  '/api/v1/access-controls/{access_control_uuid}':
    get:
      tags: ['Access Controls']
      operationId: getAccessControl
      summary: 'Lire une liste d''accès et l''intégralité de ses nœuds'
      description: |
        Renvoie la liste et **tous** ses nœuds. C'est la lecture à effectuer avant
        toute modification : `PATCH` remplace l'intégralité des nœuds, il faut donc
        partir de la liste complète (voir `PATCH`).

        Permission requise : `access_control_view`.
      parameters:
        - $ref: '#/components/parameters/AccessControlUuid'
      responses:
        200:
          description: Succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessControl'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }

    patch:
      tags: ['Access Controls']
      operationId: updateAccessControl
      summary: 'Modifier une liste d''accès — les nœuds sont REMPLACÉS EN ENTIER'
      description: |
        ## ⚠️ `nodes` est un remplacement total, jamais un ajout

        Le champ `nodes` reçoit **la liste complète et définitive** des nœuds.
        Le service supprime tous les nœuds existants de la liste, puis réinsère
        exactement ceux reçus.

        **Toute entrée omise est supprimée.** Envoyer un seul nœud pour « ajouter
        une IP » efface toutes les autres, y compris les autorisations d'un autre
        client de l'instance — les listes d'accès sont globales. La séquence
        correcte est toujours : `GET /access-controls/{access_control_uuid}`,
        modifier le tableau `nodes` reçu, puis renvoyer **ce tableau entier**.
        Envoyer `nodes: []` vide la liste.

        Un `node_cidr` que le service ne sait pas normaliser est **ignoré
        silencieusement** : il disparaît sans erreur. C'est pourquoi chaque CIDR est
        validé en amont et qu'un CIDR invalide produit un `400` plutôt qu'une liste
        amputée. Sont refusés, en plus des valeurs syntaxiquement invalides :
        `0.0.0.0/0`, les plages privées ou réservées (`10.0.0.0/8`, `172.16.0.0/12`,
        `192.168.0.0/16`, `127.0.0.0/8`, `100.64.0.0/10`) et tout masque plus large
        que `/24` sur une liste dont `access_control_default` vaut `deny`.

        ## Les attributs de la liste sont partiels, mais à envoyer complets

        `access_control_name`, `access_control_default` et
        `access_control_description` sont réécrits à chaque appel. Le contrôleur
        complète les champs absents avec les valeurs courantes de la liste, de sorte
        qu'un `PATCH` ne portant que `nodes` ne modifie ni le nom, ni le défaut, ni
        la description. Un client qui veut changer le défaut doit l'envoyer
        explicitement.

        Après écriture, le cache `acl.conf` est invalidé et `reloadacl` est émis ;
        le résultat brut est renvoyé dans `switch_response` (la réponse reste `200`,
        voir « `switch_response` » en introduction).

        Ni la création ni la suppression d'une liste ne sont exposées par cette API.

        Permission requise : `access_control_edit`.
      parameters:
        - $ref: '#/components/parameters/AccessControlUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessControlUpdateRequest'
            examples:
              remplacement_complet:
                summary: 'Liste complète des nœuds autorisés (les absents sont supprimés)'
                value:
                  nodes:
                    - node_type: allow
                      node_cidr: 192.0.2.10/32
                      node_description: 'Opérateur — SBC 1'
                    - node_type: allow
                      node_cidr: 198.51.100.0/24
                      node_description: 'Opérateur — plage média'
      responses:
        200:
          description: 'Liste modifiée ; ses nœuds sont ceux renvoyés.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessControl'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }

  # ---------------------------------------------------------------------------
  # Dialplans
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Jeton personnel (Sanctum) d'un utilisateur FS PBX, créé depuis le tableau
        de bord. Un jeton par environnement appelant. L'en-tête
        `Authorization: Bearer <token>` est obligatoire sur toutes les opérations.
  parameters:
...
    AccessControlUuid:
      name: access_control_uuid
      in: path
      required: true
      description: "UUID de la liste d'accès (globale à l'instance)."
      schema:
        type: string
        format: uuid
      example: 2b6a9f41-7c3d-4e58-9a10-3f5b8c7d6e21
...
  responses:
    BadRequest:
      description: |
        UUID de route malformé, ou corps invalide. **La validation du corps est
        rendue en `400`** (`error.code = "invalid_parameter"`, `error.param` porte
        le premier champ fautif), jamais en `422` : dans cette API, `422` est
        réservé aux erreurs du moteur téléphonique.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            uuid_malforme:
              value:
                error:
                  type: invalid_request_error
                  message: 'Invalid domain UUID.'
                  code: invalid_request
                  param: domain_uuid
                  doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
            corps_invalide:
              value:
                error:
                  type: invalid_request_error
                  message: 'Enter a valid IP address or CIDR range.'
                  code: invalid_parameter
                  param: nodes.0.node_cidr
                  doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
    Unauthorized:
      description: |
        Jeton absent, invalide ou révoqué. **Deux formes de corps** selon la couche
        qui refuse : le middleware `RequireBearerToken` (en-tête absent) renvoie
        `{"success": false, "message": "Bearer token required."}`, tandis qu'un
        jeton invalide passe par l'enveloppe d'erreur standard.
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/ErrorEnvelope'
              - $ref: '#/components/schemas/MiddlewareError'
          examples:
            entete_absente:
              summary: 'Aucun en-tête Authorization'
              value:
                success: false
                message: 'Bearer token required.'
            jeton_invalide:
              summary: 'Jeton présent mais invalide ou révoqué'
              value:
                error:
                  type: authentication_error
                  message: Unauthenticated.
                  code: unauthenticated
    Forbidden:
      description: |
        Permission manquante, ou domaine hors du périmètre de l'utilisateur.
        **Deux formes de corps** : permission manquante → forme « middleware » avec
        `error.code = "forbidden_permission"` et `error.permission` nommant la
        permission attendue ; domaine interdit → enveloppe standard avec
        `error.code = "forbidden_domain"`.
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/MiddlewareError'
              - $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            permission_manquante:
              value:
                success: false
                message: 'Forbidden (missing permission).'
                error:
                  code: forbidden_permission
                  permission: gateway_add
            domaine_interdit:
              value:
                error:
                  type: invalid_request_error
                  message: 'You do not have access to this domain.'
                  code: forbidden_domain
                  param: domain_uuid
                  doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
    NotFound:
      description: |
        Ressource inexistante, ou existante mais rattachée à un autre domaine que
        celui de la route.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              type: invalid_request_error
              message: 'Gateway not found.'
              code: resource_missing
              param: gateway_uuid
              doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
    SwitchError:
      description: |
        **Le moteur téléphonique a répondu une erreur.** `error.message` porte la
        réponse brute de FreeSWITCH, y compris la sentinelle
        `-ERR Could not connect to FreeSWITCH event socket.` quand l'event socket
        est injoignable. Aucune écriture en base n'est concernée : ces opérations
        n'en font pas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              type: invalid_request_error
              message: '-ERR Could not connect to FreeSWITCH event socket.'
              code: switch_error
              param: null
              doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
    TooManyRequests:
      description: |
        Débit dépassé : 60 requêtes par minute et par utilisateur authentifié
        (à défaut, par adresse IP). Les en-têtes `Retry-After` et
        `X-RateLimit-Remaining` accompagnent la réponse.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: |
        Erreur interne. Le message est volontairement générique ; le détail est
        journalisé côté serveur.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              type: api_error
              message: 'Internal server error.'
              code: internal_error
              doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
  schemas:
...
    ErrorEnvelope:
      type: object
      description: "Enveloppe d'erreur standard de l'API v1."
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              description: "Famille d'erreur."
              enum:
                - invalid_request_error
                - authentication_error
                - not_found
                - forbidden_method
                - api_error
            message:
              type: string
              description: "Message lisible. Pour un `422`, il contient la réponse brute du moteur téléphonique."
            code:
              type: string
              nullable: true
              description: 'Code machine, stable — c''est lui qu''un client doit tester.'
              enum:
                - unauthenticated
                - invalid_request
                - invalid_parameter
                - forbidden
                - forbidden_domain
                - forbidden_permission
                - resource_missing
                - resource_not_found
                - switch_error
                - internal_error
            param:
              type: string
              nullable: true
              description: 'Champ ou paramètre fautif, quand il est identifiable.'
            doc_url:
              type: string
              nullable: true
    MiddlewareError:
      type: object
      description: |
        Forme d'erreur renvoyée par les middlewares d'authentification et
        d'autorisation, **distincte** de l'enveloppe standard.
      required: [success, message]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          example: 'Forbidden (missing permission).'
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
              enum: [forbidden_permission]
            permission:
              type: string
              description: 'Permission attendue.'
              example: gateway_add

    # --- Passerelles ---------------------------------------------------------
...
    AccessControlNode:
      type: object
      description: 'Une entrée de liste d''accès.'
      required: [node_cidr]
      properties:
        node_type:
          type: string
          enum: [allow, deny]
          default: allow
          description: 'Toute valeur autre que `allow` ou `deny` est ramenée à `allow`.'
        node_cidr:
          type: string
          maxLength: 255
          description: |
            Adresse ou plage, en notation CIDR. Une IPv4 sans préfixe est
            normalisée en `/32`, une IPv6 sans préfixe est conservée nue.

            Un CIDR que le service ne sait pas normaliser serait **ignoré
            silencieusement** : il est donc refusé en amont (`400`). Sont
            également refusés `0.0.0.0/0`, les plages privées ou réservées
            (`10.0.0.0/8`, `172.16.0.0/12`, `192.168.0.0/16`, `127.0.0.0/8`,
            `100.64.0.0/10`) et tout masque plus large que `/24` sur une liste
            dont le défaut est `deny`.
          example: 192.0.2.10/32
        node_description:
          type: string
          nullable: true
          maxLength: 255
          example: 'Opérateur — SBC 1'
    AccessControlSummary:
      type: object
      description: "Liste d'accès sans ses nœuds."
      required: [access_control_uuid, object, access_control_name, access_control_default]
      properties:
        access_control_uuid:
          type: string
          format: uuid
          example: 2b6a9f41-7c3d-4e58-9a10-3f5b8c7d6e21
        object:
          type: string
          enum: [access_control]
        access_control_name:
          type: string
          maxLength: 255
          example: providers
        access_control_default:
          type: string
          enum: [allow, deny]
          description: 'Politique appliquée à ce qui n''est couvert par aucun nœud.'
        access_control_description:
          type: string
          nullable: true
          maxLength: 255
    AccessControl:
      allOf:
        - $ref: '#/components/schemas/AccessControlSummary'
        - type: object
          required: [nodes]
          properties:
            nodes:
              type: array
              description: "**Intégralité** des nœuds de la liste. C'est ce tableau, modifié, qu'un `PATCH` doit renvoyer."
              items:
                $ref: '#/components/schemas/AccessControlNode'
            switch_response:
              type: string
              nullable: true
              description: |
                Présent après un `PATCH` : réponse brute du moteur à `reloadacl`.
                Une valeur commençant par `-ERR` signifie que la base est à jour
                mais que FreeSWITCH applique encore l'ancienne liste. Absent (ou
                `null`) en lecture.
    AccessControlUpdateRequest:
      type: object
      description: |
        `nodes` est un **remplacement total** (voir la description de l'opération).
        Les trois autres champs sont facultatifs et conservent leur valeur courante
        s'ils sont omis.
      properties:
        access_control_name:
          type: string
          maxLength: 255
        access_control_default:
          type: string
          enum: [allow, deny]
          description: |
            Omis, ce champ conserve sa valeur courante. Le passer de `deny` à
            `allow` ouvre la liste à tout ce qu'aucun nœud ne couvre : à envoyer
            délibérément.
        access_control_description:
          type: string
          nullable: true
          maxLength: 255
        nodes:
          type: array
          description: |
            **Liste complète et définitive des nœuds.** Tout nœud absent de ce
            tableau est supprimé. `[]` vide la liste.
          items:
            $ref: '#/components/schemas/AccessControlNode'
    AccessControlListResponse:
      type: object
      description: "Enveloppe de liste de l'API v1."
      required: [object, url, has_more, data]
      properties:
        object:
          type: string
          enum: [list]
        url:
          type: string
          example: /api/v1/access-controls
        has_more:
          type: boolean
          description: 'Toujours `false` : cette collection n''est pas paginée.'
        data:
          type: array
          items:
            $ref: '#/components/schemas/AccessControlSummary'
    AccessControlReloadResponse:
      type: object
      required: [object, switch_response]
      properties:
        object:
          type: string
          enum: [access_control_reload]
        switch_response:
          type: string
          description: 'Réponse brute du moteur à `reloadacl`. Une réponse `-ERR` produit un `422`, elle n''apparaît donc jamais ici.'
          example: '+OK'

    # --- Plans de numérotation -----------------------------------------------
...